feat: add geocode save search test
